Restructure project
Make the directory hierarchy flatter.
