--- docs/source/pumaguard.cmd.rst
@@ -1,53 +1,0 @@
-pumaguard.cmd package
-=====================
-
-Submodules
-----------
-
-pumaguard.cmd.classify module
------------------------------
-
-.. automodule:: pumaguard.cmd.classify
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pumaguard.cmd.output module
----------------------------
-
-.. automodule:: pumaguard.cmd.output
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pumaguard.cmd.server module
----------------------------
-
-.. automodule:: pumaguard.cmd.server
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pumaguard.cmd.trailcam module
------------------------------
-
-.. automodule:: pumaguard.cmd.trailcam
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-pumaguard.cmd.train module
---------------------------
-
-.. automodule:: pumaguard.cmd.train
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Module contents
----------------
-
-.. automodule:: pumaguard.cmd
-   :members:
-   :undoc-members:
-   :show-inheritance:
