+++ pumaguard/trailcam.py
