+++ pyproject.toml
@@ -25,11 +25,11 @@ pytest = "^8.2.2"
 
 [tool.poetry.scripts]
 pick-files = 'scripts.pick_files:main'
-pumaguard-classify = 'pumaguard.cmd.classify:main'
-pumaguard-output = 'pumaguard.cmd.output:main'
-pumaguard-server = 'pumaguard.cmd.server:main'
-pumaguard-trailcam = 'pumaguard.cmd.trailcam:main'
-pumaguard-train = 'pumaguard.cmd.train:main'
+pumaguard-classify = 'pumaguard.classify:main'
+pumaguard-output = 'pumaguard.output:main'
+pumaguard-server = 'pumaguard.server:main'
+pumaguard-trailcam = 'pumaguard.trailcam:main'
+pumaguard-train = 'pumaguard.train:main'
 
 [build-system]
 requires = ["poetry-core"]
