+++ scripts/train.py
@@ -10,7 +10,7 @@ import os.path
 # pylint: disable=wrong-import-position
 sys.path.append(os.path.realpath(f'{os.path.dirname(__file__)}/..'))
 
-from pumaguard.cmd.train import main
+from pumaguard.train import main
 
 if __name__ == '__main__':
     sys.exit(main())
